requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
requests
python-dotenv
customtkinter>=5.2.0
keyboard>=0.13.5
pyperclip
pyttsx3
SpeechRecognition
PyAudio
configparser
google-generativeai
